+++ tests/DatasetTest.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace BakameTest\Period\Visualizer;
 
 use ArrayObject;
+use Bakame\Period\Visualizer\Dataset;
 use Bakame\Period\Visualizer\DecimalNumber;
 use Bakame\Period\Visualizer\LatinLetter;
-use Bakame\Period\Visualizer\Tuple;
 use DateTimeImmutable;
 use League\Period\Period;
 use League\Period\Sequence;
@@ -24,9 +24,9 @@ use PHPUnit\Framework\TestCase;
 use function iterator_to_array;
 
 /**
- * @coversDefaultClass \Bakame\Period\Visualizer\Tuple;
+ * @coversDefaultClass \Bakame\Period\Visualizer\Dataset;
  */
-final class TupleTest extends TestCase
+final class DatasetTest extends TestCase
 {
     public function testFromSequenceConstructor(): void
     {
@@ -34,16 +34,16 @@ final class TupleTest extends TestCase
         $periodB = Period::fromDay(2019, 3, 15);
         $labelGenerator = new LatinLetter('A');
         $sequence = new Sequence($periodA, $periodB);
-        $tuple = Tuple::fromSequence($sequence, $labelGenerator);
-        $arr = iterator_to_array($tuple);
+        $dataset = Dataset::fromSequence($sequence, $labelGenerator);
+        $arr = iterator_to_array($dataset);
 
-        self::assertCount(2, $tuple);
+        self::assertCount(2, $dataset);
         self::assertSame('B', $arr[1][0]);
         self::assertTrue($periodB->equals($arr[1][1]));
 
-        $emptyTuple = Tuple::fromSequence(new Sequence(), $labelGenerator);
-        self::assertCount(0, $emptyTuple);
-        self::assertTrue($emptyTuple->isEmpty());
+        $emptyDataset = Dataset::fromSequence(new Sequence(), $labelGenerator);
+        self::assertCount(0, $emptyDataset);
+        self::assertTrue($emptyDataset->isEmpty());
     }
 
     /**
@@ -51,10 +51,10 @@ final class TupleTest extends TestCase
      */
     public function testFromIterableConstructor(iterable $input, int $expectedCount, bool $isEmpty, bool $boundaryIsNull): void
     {
-        $tuple = Tuple::fromCollection($input);
-        self::assertCount($expectedCount, $tuple);
-        self::assertSame($isEmpty, $tuple->isEmpty());
-        self::assertSame($boundaryIsNull, null === $tuple->boundaries());
+        $dataset = Dataset::fromCollection($input);
+        self::assertCount($expectedCount, $dataset);
+        self::assertSame($isEmpty, $dataset->isEmpty());
+        self::assertSame($boundaryIsNull, null === $dataset->boundaries());
     }
 
     public function provideIterableStructure(): iterable
@@ -101,41 +101,41 @@ final class TupleTest extends TestCase
 
     public function testAppendPairs(): void
     {
-        $tuple = new Tuple([
+        $dataset = new Dataset([
             ['A', new Sequence(new Period('2018-01-01', '2018-01-15'))],
             ['B', new Sequence(new Period('2018-01-15', '2018-02-01'))],
             [new DateTimeImmutable(), new Sequence(new Period('2018-01-15', '2018-02-01'))],
             ['C', 'foo bar'],
         ]);
 
-        self::assertCount(2, $tuple);
+        self::assertCount(2, $dataset);
     }
 
     public function testLabelizePairs(): void
     {
-        $tuple = new Tuple([
+        $dataset = new Dataset([
             ['A', new Sequence(new Period('2018-01-01', '2018-01-15'))],
             ['B', new Sequence(new Period('2018-01-15', '2018-02-01'))],
         ]);
-        self::assertSame(['A', 'B'], $tuple->labels());
-        self::assertSame(1, $tuple->labelMaxLength());
+        self::assertSame(['A', 'B'], $dataset->labels());
+        self::assertSame(1, $dataset->labelMaxLength());
 
-        $newTuple = $tuple->labelize(new DecimalNumber(42));
-        self::assertSame(['42', '43'], $newTuple->labels());
-        self::assertSame($tuple->items(), $newTuple->items());
-        self::assertSame(2, $newTuple->labelMaxLength());
+        $newDataset = $dataset->labelize(new DecimalNumber(42));
+        self::assertSame(['42', '43'], $newDataset->labels());
+        self::assertSame($dataset->items(), $newDataset->items());
+        self::assertSame(2, $newDataset->labelMaxLength());
     }
 
     public function testLabelizePairsReturnsSameInstance(): void
     {
-        $tuple = new Tuple([
+        $dataset = new Dataset([
             ['A', new Sequence(new Period('2018-01-01', '2018-01-15'))],
             ['B', new Sequence(new Period('2018-01-15', '2018-02-01'))],
         ]);
 
-        self::assertEquals($tuple, $tuple->labelize(new LatinLetter()));
+        self::assertEquals($dataset, $dataset->labelize(new LatinLetter()));
 
-        $emptyTuple = new Tuple();
-        self::assertEquals($emptyTuple, $emptyTuple->labelize(new DecimalNumber(42)));
+        $emptyDataset = new Dataset();
+        self::assertEquals($emptyDataset, $emptyDataset->labelize(new DecimalNumber(42)));
     }
 }